Add files via upload
Added LoggingScript in .py and .cal. Gives the ability to call log(text) in another function. Logs the date and time and the text specified. You can use this to log when a certain change was made or to record the old values from the table somewhere before changing them.